Fix the bug in parsing a projection expression
The placeholders are not generated on when the keywords are appeared in
item's attribute name.

The expression parser classes for DynamoDB were speparated to another
module.
